# nest-api/openapi.yaml
openapi: 3.0.0
paths:
...
  /user/create-new-user:
    post:
      operationId: UserController_createNewUser
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserDto"
      responses:
        "201":
          description: ""
      summary: create-new-user
      tags: &a1
        - User
  /user/login:
    post:
      operationId: UserController_loginUser
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginUserDto"
      responses:
        "201":
          description: ""
      summary: login
      tags: *a1
...
  /user/{id}/privileges:
    patch:
      operationId: UserController_updateUserPrivileges
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RolesPrivilegesDto"
      responses:
        "200":
          description: ""
      summary: :id/privileges
      tags: *a1
...
  /price/create:
    post:
      operationId: PriceController_createProductPrice
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePriceDto"
      responses:
        "201":
          description: ""
      summary: create
      tags: *a3
...
info:
  title: 35-fit nest-api
  description: The NEST-API description for the 35-fit project
  version: "1.0"
  contact: {}
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
...
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    CreateUserDto:
      type: object
      properties: {}
    LoginUserDto:
      type: object
      properties: {}
    RolesPrivilegesDto:
      type: object
      properties: {}
...
    CreatePriceDto:
      type: object
      properties: {}
...
security:
  - bearer: []